% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_metrics.R
\name{canopy_cover_metrics}
\alias{canopy_cover_metrics}
\title{Compute Canopy Cover}
\usage{
canopy_cover_metrics(
  x,
  y,
  z,
  return_number,
  scan_angle,
  voxel_res = 3,
  angle_scale,
  density_scale
)
}
\arguments{
\item{x}{Numeric vector of X coordinates.}

\item{y}{Numeric vector of Y coordinates.}

\item{z}{Numeric vector of Z (height) values.}

\item{return_number}{Integer vector of return numbers (e.g., from LAS file).}
}
\value{
A named list with metrics including:
\itemize{
\item \code{fractional_canopy_cover}
\item \code{LAI}, \code{LAD_max}, \code{LAD_mean}, \code{LAD_z_max}
}
}
\description{
Calculates canopy cover, and leaf area density metrics
for a small area of LiDAR point cloud data.
}
