% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_decimation_weights}
\alias{compute_decimation_weights}
\title{Compute Weights for LiDAR Point Decimation}
\usage{
compute_decimation_weights(
  scan_angle,
  local_density,
  angle_scale = 1/15,
  density_scale = 1/10
)
}
\arguments{
\item{scan_angle}{Vector of scan angle values.}

\item{local_density}{Estimated local point density per voxel.}

\item{angle_scale}{Scaling factor for scan angle term (default = 1/15).}

\item{density_scale}{Scaling factor for density term (default = 1/10).}
}
\value{
A numeric vector of decimation weights between 0 and 1.
}
\description{
Computes decimation weights based on scan angle and local point density,
which are used to reduce LiDAR sampling bias.
}
