% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{connectivity_metrics_binned}
\alias{connectivity_metrics_binned}
\title{Compute Binned Graph Connectivity Metrics from LiDAR Point Cloud}
\usage{
connectivity_metrics_binned(
  x,
  y,
  z,
  edge_thresh_values = c(3, 3),
  z_1 = 1,
  z_20 = 6,
  z_40 = 12,
  voxel_res = 3,
  psid,
  QL1 = FALSE,
  n = 1,
  res = 3
)
}
\arguments{
\item{x}{X coordinates of LiDAR points.}

\item{y}{Y coordinates of LiDAR points.}

\item{z}{Z (height) values of LiDAR points.}

\item{edge_thresh_values}{Numeric vector of edge thresholds (e.g., c(3, 3)) for each vertical bin.}

\item{z_1}{Lower height threshold (default: 1 m).}

\item{z_20}{Midstory lower bound (default: 6 m).}

\item{z_40}{Midstory upper bound (default: 12 m).}

\item{voxel_res}{Numeric, voxel resolution in XY plane (default: 3).}

\item{psid}{unquoted \code{PointSourceID} from the LAS file.}

\item{QL1}{logical for whether to QL1 methods or not. Default (FALSE)}

\item{n}{Numeric to pass to \code{lidR::random_per_voxel()} density argument.}

\item{res}{Numeric to pass to \code{lidR::random_per_voxel()} res argument.}
}
\value{
A named list of graph-theoretic metrics for each bin (understory, midstory).
}
\description{
Calculates graph-theoretic metrics (e.g., betweenness, closeness, path length)
for voxelized LiDAR data within vertical height bins. This function is used to
assess 3D structural connectivity (e.g., for wildlife habitat like lynx).
}
\note{
The bounds for \verb{z_1, ,z_20, and z_40} can be whatever you want but with always be labelled \code{understory_} and \code{midstory_}.
}
