% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_metrics.R
\name{lad_metrics}
\alias{lad_metrics}
\title{Leaf Area Density Metrics}
\usage{
lad_metrics(z, dz = 1, k = 0.5, z0 = 2)
}
\arguments{
\item{z}{Numeric vector of Z heights.}

\item{dz}{Vertical resolution of profile bins (default = 1).}

\item{k}{Extinction coefficient (default = 0.5).}

\item{z0}{Minimum Z height considered vegetation (default = 2).}
}
\value{
A named list of metrics:
\itemize{
\item \code{LAI} Leaf Area Index
\item \code{LAD_max}, \code{LAD_mean}
\item \code{LAD_z_max} Height of maximum LAD
}
}
\description{
Computes LAD and LAI metrics from vertical profile of LiDAR Z values.
}
