% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{las_decimate_by_scan_and_density}
\alias{las_decimate_by_scan_and_density}
\title{Decimate LiDAR Point Cloud by Scan Angle and Local Density}
\usage{
las_decimate_by_scan_and_density(
  las,
  voxel_res = 3,
  angle_scale = 1/15,
  density_scale = 1/10
)
}
\arguments{
\item{las}{A LAS object from the \code{lidR} package.}

\item{voxel_res}{Resolution of voxel used for density estimation (default = 3).}

\item{angle_scale}{Scaling factor for decimating by scan angle.}

\item{density_scale}{Scaling factor for decimating by local density.}
}
\value{
A decimated LAS object.
}
\description{
Applies spatial decimation to a LAS object based on both scan angle and
local voxel point density, reducing sampling artifacts.
}
