% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_graph_deps}
\alias{load_graph_deps}
\title{Load Required Graph and LiDAR Dependencies}
\usage{
load_graph_deps()
}
\value{
Invisibly returns \code{TRUE} if all packages are successfully loaded.
}
\description{
Loads necessary packages used in tree detection and voxel-based graph computation.
Suppresses loading messages.
}
