% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_zero_metrics}
\alias{named_zero_metrics}
\title{Return a List of Named NA Metrics for Zero-Point Tiles}
\usage{
named_zero_metrics(type = "graph")
}
\arguments{
\item{type}{Character string, either \code{"graph"}, \code{"canopy"} or \code{"trees"}. Determines which metric structure to return.}
}
\value{
A named list of metrics filled with \code{NA_real_} values appropriate to the type.
}
\description{
Provides default NA values for graph or canopy metrics when input data is insufficient.
}
