% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_metrics.R
\name{canopy_cover_metrics}
\alias{canopy_cover_metrics}
\title{Compute Canopy Cover}
\usage{
canopy_cover_metrics(x, y, z, psid, QL1 = FALSE, n = 15, res = 3, ...)
}
\arguments{
\item{x}{Numeric vector of X coordinates.}

\item{y}{Numeric vector of Y coordinates.}

\item{z}{Numeric vector of Z (height) values.}

\item{psid}{unquoted \code{PointSourceID} from the LAS file.}

\item{QL1}{logical for whether to QL1 methods or not. Default (FALSE)}

\item{...}{Arguments to pass to \code{lidar.metrics::tree_detection()}.}

\item{density}{Numeric to pass to \code{lidR::homogenize()} density argument.}

\item{dec_res}{Numeric to pass to \code{lidR::homogenize()} res argument.}
}
\value{
A named list with metrics including:
\itemize{
\item \code{fractional_canopy_cover}
\item \code{rumple_index}
\item \code{LAI}, \code{LAD_max}, \code{LAD_mean}, \code{LAD_z_max}
}
}
\description{
Calculates canopy cover, and leaf area density metrics
for a small area of LiDAR point cloud data.
}
