% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{compute_graph_metrics}
\alias{compute_graph_metrics}
\title{Compute Graph Metrics for a Voxelized Vertical Bin}
\usage{
compute_graph_metrics(las, z_min, z_max, edge_thresh, voxel_res, QL1, n, res)
}
\arguments{
\item{las}{A LAS object (LiDAR point cloud) from the \code{lidR} package.}

\item{z_min}{Minimum Z height to include in bin.}

\item{z_max}{Maximum Z height to include in bin.}

\item{edge_thresh}{Distance threshold (in meters) for connecting centroids.}

\item{voxel_res}{Numeric voxel resolution for binning.}

\item{QL1}{logical for whether to QL1 methods or not.}

\item{n}{Numeric to pass to \code{lidR::random_per_voxel()} density argument.}

\item{res}{Numeric to pass to \code{lidR::random_per_voxel()} res argument.}
}
\value{
A named list of graph metrics.
}
\description{
Calculates key graph-theoretic metrics for a filtered LAS object, including
strength, degree, betweenness, closeness, and more. Uses convex hull volume
as edge weights.
}
