% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_metric_lynx.R
\name{compute_graph_metrics}
\alias{compute_graph_metrics}
\title{Compute Graph Metrics for a Filtered LAS Within a Vertical Bin}
\usage{
compute_graph_metrics(las, z_min, z_max, edge_thresh, voxel_res)
}
\arguments{
\item{las}{A LAS object (point cloud) from the \code{lidR} package.}

\item{z_min}{Minimum Z height threshold.}

\item{z_max}{Maximum Z height threshold.}

\item{edge_thresh}{Numeric, distance threshold for creating edges.}

\item{voxel_res}{Numeric, XY voxel resolution.}
}
\value{
A named list of graph-theoretic metrics.
}
\description{
Creates a voxel-based graph using a fixed edge threshold, and calculates graph
metrics such as weighted degree, betweenness, closeness, and component structure.
}
